refactor: move SignInController and SignUpController to authentication folder
