+++ src/main/factories/controllers/account/sign-in-controller-factory.ts
@@ -1,7 +1,7 @@
 import { AuthenticationFactory } from '@/main/factories/usecases/account/commands'
 import { LogErrorControllerDecoratorFactory } from '@/main/factories/decorators/log'
 import { type Controller } from '@/presentation/protocols'
-import { SignInController } from '@/presentation/controllers/account'
+import { SignInController } from '@/presentation/controllers/account/authentication'
 
 export class SignInControllerFactory {
   public static readonly makeSignInController = (): Controller => {
