+++ tests/presentation/controllers/account/authentication/sign-up-controller.spec.ts
@@ -3,7 +3,7 @@ import { faker } from '@faker-js/faker'
 import { throwError } from '@/tests/test-helper'
 import { AddAccountSpy, AuthenticationSpy } from '@/tests/presentation/mocks/account'
 import { mockAddAccountInput } from '@/tests/domain/mocks/account'
-import { SignUpController } from '@/presentation/controllers/account'
+import { SignUpController } from '@/presentation/controllers/account/authentication'
 import { HttpHelper } from '@/presentation/helpers'
 import { EntityValidationError } from '@/domain/errors'
 import { EmailInUseError, AccountNotFoundError, InvalidCredentialsError } from '@/domain/errors/account'
